fix Vector2/3b type casts

+++ Automata/Numerics/Vector2b.Static.cs
@@ -22,7 +22,7 @@ namespace Automata.Numerics
         {
             if (Sse2.IsSupported)
             {
-                return (Vector2b)Sse2.CompareEqual((Vector128<int>)a, (Vector128<int>)b);
+                return (Vector2b)Sse2.CompareEqual((Vector128<byte>)a, (Vector128<byte>)b);
             }
             else
             {
